refactor(JOE-942): extract work-store retention + storage lock modules
Move runWorkStoreRetentionMaintenance, receipt/run/audit ledger prune
helpers, and pruneWorkEvents into work-store/retention.ts. Extract
assertNoStorageOperationInProgress to work-store/storage-lock.ts.

Façade drops ~310 LOC (5693 → 5382). Public API re-exported unchanged.
Retention vitest suite green.

Refs JOE-919, JOE-942
